requirements.txt
nibabel>=5.0,<6.0
pydicom>=2.4,<4.0
numpy>=1.24,<3.0
scipy>=1.10,<2.0
pandas>=2.0,<3.0
matplotlib>=3.7,<4.0
antspyx>=0.5,<1.0
